az-mlops-mvp/backend/requirements.txt

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

pydantic==2.10.4
pydantic-settings==2.7.0

# Databricks SDK
databricks-sdk==0.39.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Environment management
python-dotenv==1.0.1

# HTTP client (used internally by SDK + testing)
httpx==0.28.1

pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
anyio==4.7.0

python-dateutil==2.9.0
